Refactor project type serializers to the their own file
